// cluster/portal/portal/web/package/src/apis/metav1/metav1.ts
export interface ObjectReference {
    /**
     * APIVersion is the API version of the reference resource
     *
     * @generated from protobuf field: string apiVersion = 1
     */
    apiVersion: string;
    /**
     * Kind is the resource kind. A kind is unique per an APIVersion.
     *
     * @generated from protobuf field: string kind = 2
     */
    kind: string;
    /**
     * UID is the resource object UID. A UID is unique across the entire Cluster.
     *
     * @generated from protobuf field: string uid = 3
     */
    uid: string;
    /**
     * Name is the resource object name. A name is unique per a certain kind.
     *
     * @generated from protobuf field: string name = 4
     */
    name: string;
    /**
     * ResourceVersion is the resourceVersion of the resource.
     *
     * @generated from protobuf field: string resourceVersion = 5
     */
    resourceVersion: string;
}
// ...
export interface GetOptions {
    /**
     * UID is the object's UID
     *
     * @generated from protobuf field: string uid = 1
     */
    uid: string;
    /**
     * Name is the object's name
     *
     * @generated from protobuf field: string name = 2
     */
    name: string;
}
// ...
class ObjectReference$Type extends MessageType<ObjectReference> {
    constructor() {
        super("octelium.api.main.meta.v1.ObjectReference", [
            { no: 1, name: "apiVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "resourceVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ObjectReference>): ObjectReference {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.apiVersion = "";
        message.kind = "";
        message.uid = "";
        message.name = "";
        message.resourceVersion = "";
        if (value !== undefined)
            reflectionMergePartial<ObjectReference>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ObjectReference): ObjectReference {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string apiVersion */ 1:
                    message.apiVersion = reader.string();
                    break;
                case /* string kind */ 2:
                    message.kind = reader.string();
                    break;
                case /* string uid */ 3:
                    message.uid = reader.string();
                    break;
                case /* string name */ 4:
                    message.name = reader.string();
                    break;
                case /* string resourceVersion */ 5:
                    message.resourceVersion = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ObjectReference, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string apiVersion = 1; */
        if (message.apiVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.apiVersion);
        /* string kind = 2; */
        if (message.kind !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.kind);
        /* string uid = 3; */
        if (message.uid !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.uid);
        /* string name = 4; */
        if (message.name !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.name);
        /* string resourceVersion = 5; */
        if (message.resourceVersion !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.resourceVersion);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.ObjectReference
 */
export const ObjectReference = new ObjectReference$Type();
// ...
class GetOptions$Type extends MessageType<GetOptions> {
    constructor() {
        super("octelium.api.main.meta.v1.GetOptions", [
            { no: 1, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<GetOptions>): GetOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.uid = "";
        message.name = "";
        if (value !== undefined)
            reflectionMergePartial<GetOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetOptions): GetOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string uid */ 1:
                    message.uid = reader.string();
                    break;
                case /* string name */ 2:
                    message.name = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string uid = 1; */
        if (message.uid !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.uid);
        /* string name = 2; */
        if (message.name !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.name);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.GetOptions
 */
export const GetOptions = new GetOptions$Type();